+++ examples/resources/cpanel_mysql_remote_host/import.sh
@@ -1,0 +1,1 @@
+terraform import cpanel_mysql_remote_host.reporting 198.51.100.20